TextArea.js moves, style components relocated and imported, all relevant imports altered
